新增删除token前判断token是否相同

+++ src/main/java/com/codeisright/attendance/security/JwtLogoutSuccessHandler.java
@@ -8,6 +8,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.logout.SimpleUrlLogoutSuccessHandler;
 
+import java.io.IOException;
+
 public class JwtLogoutSuccessHandler extends SimpleUrlLogoutSuccessHandler {
     private final UserDetailsServiceImpl userDetailsService;
 
@@ -16,29 +18,31 @@ public class JwtLogoutSuccessHandler extends SimpleUrlLogoutSuccessHandler {
     }
 
     @Override
-    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication){
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response,
+                                Authentication authentication) throws IOException {
         String token = request.getHeader("Authorization").substring(7);
         Claims claims = Jwts.parser()
                 .setSigningKey("secret".getBytes())
                 .parseClaimsJws(token)
                 .getBody();
         String Id = claims.getSubject();
+        String currentToken = userDetailsService.getJwt(Id);
+        if (!token.equals(currentToken)) {
+            logger.error("Invalid token");
+            response.getWriter().write("{\"userId\":\"" + Id + "\",\"status\":\"error\",\"msg\":\"Logout failed. " +
+                    "Wrong token.\"}");
+            return;
+        }
+        logger.info("User: " + Id + " logout successfully.");
         userDetailsService.deleteJwt(Id);
 
         // 返回一个json
         response.setContentType("application/json;charset=utf-8");
         response.setStatus(HttpServletResponse.SC_OK);
-        if(userDetailsService.getJwt(Id)==null)
-            try {
-                response.getWriter().write("{\"userId\":\"" + Id + "\",\"status\":\"ok\",\"msg\":\"Logout success\"}");
-            } catch (Exception e) {
-                logger.error("Error happened when writing response" + e);
-            }
+        if (userDetailsService.getJwt(Id) == null)
+            response.getWriter().write("{\"userId\":\"" + Id + "\",\"status\":\"ok\",\"msg\":\"Logout success\"}");
         else
-            try {
-                response.getWriter().write("{\"userId\":\"" + Id + "\",\"status\":\"error\",\"msg\":\"Logout failed. Please try again\"}");
-            }catch (Exception e) {
-                logger.error("Error happened when writing response" + e);
-            }
+            response.getWriter().write("{\"userId\":\"" + Id + "\",\"status\":\"error\",\"msg\":\"Logout failed. " +
+                    "Please try again\"}");
     }
 }
